+++ myjobs/src/main/java/com/dimogo/open/myjobs/dto/RuntimeInfo.java
@@ -275,7 +275,7 @@ public class RuntimeInfo {
 		for (DiskRuntimeInfo disk : disks) {
 			usedPercent += disk.getUsedPercent();
 		}
-		return usedPercent;
+		return usedPercent * 100;
 	}
 
 	public RuntimeInfo() {
@@ -376,456 +376,6 @@ public class RuntimeInfo {
 			}
 		} catch (Throwable e) {
 			e.printStackTrace();
-		}
-	}
-
-	private static class EthernetRuntimeInfo {
-		private String device;
-		private String ip;
-		private String broadcast;
-		private String mac;
-		private String mask;
-		private String description;
-		private String type;
-
-		public EthernetRuntimeInfo() {
-
-		}
-
-		private EthernetRuntimeInfo(NetInterfaceConfig config) {
-			device = config.getName();
-			ip = config.getAddress();
-			broadcast = config.getBroadcast();
-			mac = config.getHwaddr();
-			mask = config.getNetmask();
-			description = config.getDescription();
-			type = config.getType();
-		}
-
-		public void setDevice(String device) {
-			this.device = device;
-		}
-
-		public void setIp(String ip) {
-			this.ip = ip;
-		}
-
-		public void setBroadcast(String broadcast) {
-			this.broadcast = broadcast;
-		}
-
-		public void setMac(String mac) {
-			this.mac = mac;
-		}
-
-		public void setMask(String mask) {
-			this.mask = mask;
-		}
-
-		public void setDescription(String description) {
-			this.description = description;
-		}
-
-		public void setType(String type) {
-			this.type = type;
-		}
-
-		public String getDevice() {
-			return device;
-		}
-
-		public String getIp() {
-			return ip;
-		}
-
-		public String getBroadcast() {
-			return broadcast;
-		}
-
-		public String getMac() {
-			return mac;
-		}
-
-		public String getMask() {
-			return mask;
-		}
-
-		public String getDescription() {
-			return description;
-		}
-
-		public String getType() {
-			return type;
-		}
-	}
-
-	private static class NetRuntimeInfo {
-		private String device;
-		private String ip;
-		private String mask;
-
-		private long rxPackets;
-		private long txPackets;
-		private long rxBytes;
-		private long txBytes;
-		private long rxErrors;
-		private long txErrors;
-		private long rxDropped;
-		private long txDropped;
-
-		public NetRuntimeInfo() {
-
-		}
-
-		private NetRuntimeInfo(NetInterfaceConfig config, NetInterfaceStat stat) {
-			device = config.getName();
-			ip = config.getAddress();
-			mask = config.getNetmask();
-
-			if (stat != null) {
-				rxPackets = stat.getRxPackets();
-				txPackets = stat.getTxPackets();
-				rxBytes = stat.getRxBytes();
-				txBytes = stat.getTxBytes();
-				rxErrors = stat.getRxErrors();
-				txErrors = stat.getTxErrors();
-				rxDropped = stat.getRxDropped();
-				txDropped = stat.getTxDropped();
-			}
-		}
-
-		public void setDevice(String device) {
-			this.device = device;
-		}
-
-		public void setIp(String ip) {
-			this.ip = ip;
-		}
-
-		public void setMask(String mask) {
-			this.mask = mask;
-		}
-
-		public void setRxPackets(long rxPackets) {
-			this.rxPackets = rxPackets;
-		}
-
-		public void setTxPackets(long txPackets) {
-			this.txPackets = txPackets;
-		}
-
-		public void setRxBytes(long rxBytes) {
-			this.rxBytes = rxBytes;
-		}
-
-		public void setTxBytes(long txBytes) {
-			this.txBytes = txBytes;
-		}
-
-		public void setRxErrors(long rxErrors) {
-			this.rxErrors = rxErrors;
-		}
-
-		public void setTxErrors(long txErrors) {
-			this.txErrors = txErrors;
-		}
-
-		public void setRxDropped(long rxDropped) {
-			this.rxDropped = rxDropped;
-		}
-
-		public void setTxDropped(long txDropped) {
-			this.txDropped = txDropped;
-		}
-
-		public String getDevice() {
-			return device;
-		}
-
-		public String getIp() {
-			return ip;
-		}
-
-		public String getMask() {
-			return mask;
-		}
-
-		public long getRxPackets() {
-			return rxPackets;
-		}
-
-		public long getTxPackets() {
-			return txPackets;
-		}
-
-		public long getRxBytes() {
-			return rxBytes;
-		}
-
-		public long getTxBytes() {
-			return txBytes;
-		}
-
-		public long getRxErrors() {
-			return rxErrors;
-		}
-
-		public long getTxErrors() {
-			return txErrors;
-		}
-
-		public long getRxDropped() {
-			return rxDropped;
-		}
-
-		public long getTxDropped() {
-			return txDropped;
-		}
-	}
-
-	private static class DiskRuntimeInfo {
-		private String device;
-		private String dir;
-		private long flags;
-		private String fs;
-		private String type;
-		private long reads;
-		private long writes;
-
-		private long total;
-		private long free;
-		private long avil;
-		private long used;
-		private double usedPercent;
-
-		public DiskRuntimeInfo() {
-
-		}
-
-		private DiskRuntimeInfo(FileSystem fs, FileSystemUsage usage) {
-			device = fs.getDevName();
-			dir = fs.getDirName();
-			flags = fs.getFlags();
-			this.fs = fs.getSysTypeName();
-			type = fs.getTypeName();
-			reads = usage.getDiskReads();
-			writes = usage.getDiskWrites();
-
-			if (fs.getType() == 2) {
-				total = usage.getTotal();
-				free = usage.getFree();
-				avil = usage.getAvail();
-				used = usage.getUsed();
-				usedPercent = usage.getUsePercent();
-			}
-		}
-
-		public void setDevice(String device) {
-			this.device = device;
-		}
-
-		public void setDir(String dir) {
-			this.dir = dir;
-		}
-
-		public void setFlags(long flags) {
-			this.flags = flags;
-		}
-
-		public void setFs(String fs) {
-			this.fs = fs;
-		}
-
-		public void setType(String type) {
-			this.type = type;
-		}
-
-		public void setReads(long reads) {
-			this.reads = reads;
-		}
-
-		public void setWrites(long writes) {
-			this.writes = writes;
-		}
-
-		public void setTotal(long total) {
-			this.total = total;
-		}
-
-		public void setFree(long free) {
-			this.free = free;
-		}
-
-		public void setAvil(long avil) {
-			this.avil = avil;
-		}
-
-		public void setUsed(long used) {
-			this.used = used;
-		}
-
-		public void setUsedPercent(double usedPercent) {
-			this.usedPercent = usedPercent;
-		}
-
-		public double getUsedPercent() {
-			return usedPercent;
-		}
-
-		public String getDevice() {
-			return device;
-		}
-
-		public String getDir() {
-			return dir;
-		}
-
-		public long getFlags() {
-			return flags;
-		}
-
-		public String getFs() {
-			return fs;
-		}
-
-		public String getType() {
-			return type;
-		}
-
-		public long getReads() {
-			return reads;
-		}
-
-		public long getWrites() {
-			return writes;
-		}
-
-		public long getTotal() {
-			return total;
-		}
-
-		public long getFree() {
-			return free;
-		}
-
-		public long getAvil() {
-			return avil;
-		}
-
-		public long getUsed() {
-			return used;
-		}
-	}
-
-	private static class CpuRuntimeInfo {
-		private int mHz;
-		private String vendor;
-		private String model;
-		private long cacheSize;
-
-		private double user;
-		private double sys;
-		private double wait;
-		private double nice;
-		private double idle;
-		private double combined;
-
-		public CpuRuntimeInfo() {
-
-		}
-
-		private CpuRuntimeInfo(CpuInfo info, CpuPerc perc) {
-			mHz = info.getMhz();
-			vendor = info.getVendor();
-			model = info.getModel();
-			cacheSize = info.getCacheSize();
-
-			user = perc.getUser();
-			sys = perc.getSys();
-			wait = perc.getWait();
-			nice = perc.getNice();
-			idle = perc.getIdle();
-			combined = perc.getCombined();
-		}
-
-		public void setmHz(int mHz) {
-			this.mHz = mHz;
-		}
-
-		public void setVendor(String vendor) {
-			this.vendor = vendor;
-		}
-
-		public void setModel(String model) {
-			this.model = model;
-		}
-
-		public void setCacheSize(long cacheSize) {
-			this.cacheSize = cacheSize;
-		}
-
-		public void setUser(double user) {
-			this.user = user;
-		}
-
-		public void setSys(double sys) {
-			this.sys = sys;
-		}
-
-		public void setWait(double wait) {
-			this.wait = wait;
-		}
-
-		public void setNice(double nice) {
-			this.nice = nice;
-		}
-
-		public void setIdle(double idle) {
-			this.idle = idle;
-		}
-
-		public void setCombined(double combined) {
-			this.combined = combined;
-		}
-
-		public int getmHz() {
-			return mHz;
-		}
-
-		public String getVendor() {
-			return vendor;
-		}
-
-		public String getModel() {
-			return model;
-		}
-
-		public long getCacheSize() {
-			return cacheSize;
-		}
-
-		public double getUser() {
-			return user;
-		}
-
-		public double getSys() {
-			return sys;
-		}
-
-		public double getWait() {
-			return wait;
-		}
-
-		public double getNice() {
-			return nice;
-		}
-
-		public double getIdle() {
-			return idle;
-		}
-
-		public double getCombined() {
-			return combined;
 		}
 	}
 }
